move MethodUsageResult to activation.py
