+++ tests/unit/test_core/test_activationresult.py
@@ -3,9 +3,8 @@ from unittest.mock import Mock
 
 import pytest
 
-from wakepy.core import ActivationResult
+from wakepy.core import ActivationResult, MethodUsageResult
 from wakepy.core.activationresult import (
-    MethodUsageResult,
     ModeSwitcher,
     StageName,
     UsageStatus,
