+++ wakepy/core/__init__.py
@@ -4,6 +4,7 @@ as implementation details!
 See the public Python API at: https://wakepy.readthedocs.io/
 """
 
+from .activation import MethodUsageResult as MethodUsageResult
 from .activationresult import ActivationResult as ActivationResult
 from .calls import Call as Call
 from .calls import DbusMethodCall as DbusMethodCall
